+++ src/core/targeting/build-page-targeting.spec.ts
@@ -2,8 +2,8 @@ import { cmp as cmp_ } from '@guardian/consent-management-platform';
 import type { ConsentState } from '@guardian/consent-management-platform/dist/types';
 import type { TCFv2ConsentState } from '@guardian/consent-management-platform/dist/types/tcfv2';
 import { setCookie, storage } from '@guardian/libs';
-import { getAuthStatus as getAuthStatus_ } from 'identity/api';
-import type { AuthStatus } from 'identity/api';
+import { getAuthStatus as getAuthStatus_ } from 'lib/identity/api';
+import type { AuthStatus } from 'lib/identity/api';
 import { getLocale as getLocale_ } from '../lib/get-locale';
 import type { Edition } from '../types';
 import { buildPageTargeting } from './build-page-targeting';
@@ -31,7 +31,7 @@ jest.mock('@guardian/consent-management-platform', () => ({
 	},
 }));
 
-jest.mock('identity/api', () => ({
+jest.mock('lib/identity/api', () => ({
 	isUserLoggedInOktaRefactor: () => true,
 	getAuthStatus: jest.fn(),
 	getOptionsHeadersWithOkta: jest.fn(),