+++ src/types/global.ts
@@ -7,7 +7,7 @@ import type {
 	GoogleTrackConversionObject,
 	NetworkInformation,
 } from '../core/types';
-import type { DfpEnv } from '../dfp/dfp-env';
+import type { DfpEnv } from '../lib/dfp/dfp-env';
 import type { IasPETSlot } from './ias';
 
 type ServerSideABTest = `${string}${'Variant' | 'Control'}`;
@@ -432,6 +432,9 @@ interface HeaderNotification {
 declare global {
 	interface Navigator {
 		readonly connection?: NetworkInformation;
+		readonly cookieDeprecationLabel?: {
+			getValue: () => Promise<string>;
+		};
 	}
 	interface Window {
 		guardian: {